Refactor oidcconfig and move a reuse-able code to a new helper packagex
